Split neet.core into neet.interfaces, neet.paths
