+++ neet/__init__.py
@@ -1,5 +1,6 @@
 # Copyright 2017 ELIFE. All rights reserved.
 # Use of this source code is governed by a MIT
 # license that can be found in the LICENSE file.
-from .core import *
+from .interfaces import *
 from .states import *
+from .paths import *
