+++ test/test_paths.py
@@ -10,46 +10,8 @@ class TestCore(unittest.TestCase):
         def update(self, lattice):
             pass
 
-    class FixedSizeNetwork(IsNetwork):
-        def size(self):
-            return 5
-
     class IsNotNetwork(object):
         pass
-
-    class NotFixedSizedNetwork(IsNotNetwork):
-        def size(self):
-            return 5
-
-    def test_is_network(self):
-        net = self.IsNetwork()
-        self.assertTrue(neet.is_network(net))
-        self.assertTrue(neet.is_network(type(net)))
-
-        not_net = self.IsNotNetwork()
-        self.assertFalse(neet.is_network(not_net))
-        self.assertFalse(neet.is_network(type(not_net)))
-
-        self.assertFalse(neet.is_network(5))
-        self.assertFalse(neet.is_network(int))
-
-    def test_is_fixed_sized(self):
-        net = self.IsNetwork()
-        self.assertFalse(neet.is_fixed_sized(net))
-        self.assertFalse(neet.is_fixed_sized(type(net)))
-
-        not_net = self.IsNotNetwork()
-        self.assertFalse(neet.is_fixed_sized(not_net))
-        self.assertFalse(neet.is_fixed_sized(type(not_net)))
-
-        net = self.FixedSizeNetwork()
-        self.assertTrue(neet.is_fixed_sized(net))
-        self.assertTrue(neet.is_fixed_sized(type(net)))
-
-        not_net = self.NotFixedSizedNetwork()
-        self.assertFalse(neet.is_fixed_sized(not_net))
-        self.assertFalse(neet.is_fixed_sized(type(not_net)))
-
 
     def test_trajectory_not_network(self):
         with self.assertRaises(TypeError):
